mysten_common::sync

Module notify_read

Structs

NotifyRead

Registration

Registration resolves to the value but also provides safe cancellation When Registration is dropped before it is resolved, we de-register from the pending list

Constants

CHECKPOINT_BUILDER_NOTIFY_READ_TASK_NAME