Struct mysten_common::sync::notify_read::Registration

pub struct Registration<'a, K: Eq + Hash + Clone, V: Clone> { /* private fields */ }

Registration resolves to the value but also provides safe cancellation When Registration is dropped before it is resolved, we de-register from the pending list